fixed some errors in descriptor files

+++ eclipse-workspace/MINT64/02.Kernel64/Source/Descriptor.h
@@ -7,22 +7,22 @@
  * macros that defines each flag in global descriptor
  */
 
-#define GDT_TYPE_CODE			0x0A
-#define GDT_TYPE_DATA			0x02
-#define GDT_TYPE_TSS				0x09
+#define GDT_TYPE_CODE					0x0A
+#define GDT_TYPE_DATA					0x02
+#define GDT_TYPE_TSS						0x09
 
 // s field in entry mean system.
-#define GDT_FLAGS_LOWER_S		0x10
+#define GDT_FLAGS_LOWER_S				0x10
 #define GDT_FLAGS_LOWER_DPL0			0x00
 #define GDT_FLAGS_LOWER_DPL1			0x20
 #define GDT_FLAGS_LOWER_DPL2			0x40
 #define GDT_FLAGS_LOWER_DPL3			0x60
-#define GDT_FLAGS_LOWER_P		0x80
+#define GDT_FLAGS_LOWER_P				0x80
 
-#define GDT_FLAGS_UPPER_L		0x20
+#define GDT_FLAGS_UPPER_L				0x20
 // this is D field which decides operation size
-#define GDT_FLAGS_UPPER_DB		0x40
-#define GDT_FLAGS_UPPER_G		0x80
+#define GDT_FLAGS_UPPER_DB				0x40
+#define GDT_FLAGS_UPPER_G				0x80
 
 // macros that we use in Descriptor.c
 
@@ -51,54 +51,54 @@
 #define GDT_FLAGS_UPPER_TSS (GDT_FLAGS_UPPER_G)
 
 // set descriptor offset in GDT
-#define GDT_KERNELCODESEGMENT	0x08
-#define GDT_KERNELDATASEGMENT	0x10
-#define GDT_TSSSEGMENT			0x18
+#define GDT_KERNELCODESEGMENT			0x08
+#define GDT_KERNELDATASEGMENT			0x10
+#define GDT_TSSSEGMENT					0x18
 
 // GDTR start address is 1MB + 264KB page table
-#define GDTR_STARTADDRESS		0x142000
+#define GDTR_STARTADDRESS				0x142000
 // number of 8 byte entry; null descriptor/ kernel code/ kernel data
-#define GDT_MAXENTRY8COUNT		3
+#define GDT_MAXENTRY8COUNT				3
 // number of 16 byte entry; TSS entry
-#define GDT_MAXENTRY16COUNT		1
+#define GDT_MAXENTRY16COUNT				1
 // size of Global Descriptor table
-#define GDT_TABLESIZE			( ( sizeof( GDTENTRY8 ) * GDT_MAXENTRY8COUNT ) \
+#define GDT_TABLESIZE					( ( sizeof( GDTENTRY8 ) * GDT_MAXENTRY8COUNT ) \
 		+ ( sizeof( GDTENTRY16 ) * GDT_MAXENTRY16COUNT ) )
-#define TSS_SEGMENTSIZE			( sizeof( TSSSEGMENT ) )
+#define TSS_SEGMENTSIZE					( sizeof( TSSSEGMENT ) )
 
 /*
  * macros about Interrupt Descriptor Table
  */
 
 // flags and types of Interrupt Descriptor
-#define IDT_TYPE_INTERRUPT		0x0E
-#define IDT_TYPE_TRAP			0x0F
-#define IDT_FLAGS_DPL0			0x00
-#define IDT_FLAGS_DPL1			0x20
-#define IDT_FLAGS_DPL2			0x40
-#define IDT_FLAGS_DPL3			0x60
-#define IDT_FLAGS_P				0x80
-#define IDT_FLAGS_IST0			0
-#define IDT_FLAGS_IST1			1
+#define IDT_TYPE_INTERRUPT				0x0E
+#define IDT_TYPE_TRAP					0x0F
+#define IDT_FLAGS_DPL0					0x00
+#define IDT_FLAGS_DPL1					0x20
+#define IDT_FLAGS_DPL2					0x40
+#define IDT_FLAGS_DPL3					0x60
+#define IDT_FLAGS_P						0x80
+#define IDT_FLAGS_IST0					0
+#define IDT_FLAGS_IST1					1
 
 // flags that I use for Mint64OS Interrupt Descriptor
-#define IDT_FLAGS_KERNEL			( IDT_FLAGS_DPL0 | IDT_FLAGS_P )
-#define IDT_FLAGS_USER			( IDT_FLAGS_DPL3 | IDT_FLAGS_P )
+#define IDT_FLAGS_KERNEL					( IDT_FLAGS_DPL0 | IDT_FLAGS_P )
+#define IDT_FLAGS_USER					( IDT_FLAGS_DPL3 | IDT_FLAGS_P )
 
 // maximum number of IDT entries are 100 for CPU
-#define IDT_MAXENTRYCOUNT		100
+#define IDT_MAXENTRYCOUNT				100
 // start address oF IDTR; arranged after TSS Segment
-#define IDTR_STARTADDRESS		( GDTR_STARTADDRESS + sizeof( GDTR ) + \
+#define IDTR_STARTADDRESS				( GDTR_STARTADDRESS + sizeof( GDTR ) + \
 		GDT_TABLESIZE + TSS_SEGMENTSIZE )
 // start address of Interrupt Descriptor Table
-#define IDT_STARTADDRESS			( IDTR_STARTADDRESS + sizeof( IDTR ) )
+#define IDT_STARTADDRESS					( IDTR_STARTADDRESS + sizeof( IDTR ) )
 // size of Interrupt Descriptor Table
-#define IDT_TABLESIZE			( IDT_MAXENTRYCOUNT * sizeof( IDTENTRY ) )
+#define IDT_TABLESIZE					( IDT_MAXENTRYCOUNT * sizeof( IDTENTRY ) )
 
 // IST start address
-#define IST_STARTADDRESS			0x700000
+#define IST_STARTADDRESS					0x700000
 // IST size
-#define IST_SIZE					0x100000
+#define IST_SIZE							0x100000
 
 #pragma pack( push, 1 )
 
@@ -157,7 +157,7 @@ typedef struct kIDTEntryStruct {
 	BYTE bTypeAndFlags;
 	WORD wMiddleBaseAddress;
 	DWORD dwUpperBaseAddress;
-	WORD dwReserved;
+	DWORD dwReserved;
 } IDTENTRY;
 
 #pragma pack ( pop )
